moved createTemplate functionality and template files from cli to cli-lib
