+++ packages/cli-lib/defaults/blog-listing-template.html
