+++ packages/cli-lib/defaults/blog-post-template.html
