+++ packages/cli-lib/defaults/email-template.html
