+++ packages/cli-lib/defaults/global-partial.html
@@ -1,0 +1,8 @@
+<!--
+    templateType: global_partial
+    label: Global partial
+    isAvailableForNewContent: false
+-->
+<!-- Begin partial -->
+
+<!-- End partial -->
