+++ packages/cli-lib/defaults/search-template.html
