--- packages/cli/defaults/sample-section.html
@@ -1,15 +1,0 @@
-<!--
-  templateType: section
-  isAvailableForNewContent: false
--->
-{% dnd_section %}
-  {% dnd_column %}
-    {% dnd_row %}
-      {% dnd_module
-        path='@hubspot/rich_text',
-        html='<h1>Hello world</h1>'
-      %}
-      {% end_dnd_module %}
-    {% end_dnd_row %}
-  {% end_dnd_column %}
-{% end_dnd_section %}
